Changing variable id_ from int to bigint in log_trigger

+++ lib/pytis/dbdefs/sql/log_trigger.sql
@@ -1,5 +1,5 @@
 declare
-  id_ int;
+  id_ bigint;
   current_record record;
   c text;
   cc text;
